@@ -3,9 +3,10 @@
 buildscript {
     repositories {
         jcenter()
+        google()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:2.1.3'
+        classpath 'com.android.tools.build:gradle:3.6.1'
         classpath 'com.github.dcendents:android-maven-gradle-plugin:1.4.1'
 
         // NOTE: Do not place your application dependencies here; they belong
@@ -16,6 +17,7 @@ buildscript {
 allprojects {
     repositories {
         jcenter()
+        google()
     }
 }
 
@@ -25,16 +27,16 @@ task clean(type: Delete) {
 
 ext {
     minSdkVersion = 15
-    compileSdkVersion = 24
+    compileSdkVersion = 28
     targetSdkVersion = 24
     buildToolsVersion = "23.0.2"
 
     // android
-    appcompat = 'com.android.support:appcompat-v7:24.1.1'
+    appcompat = 'androidx.appcompat:appcompat:1.0.0'
     //supportv4 = 'com.android.support:support-v4:24.1.1'
     //supportannotations = 'com.android.support:support-annotations:24.1.1'
     //recyclerview = 'com.android.support:recyclerview-v7:24.1.1'
-    design = 'com.android.support:design:24.1.1'
+    design = 'com.google.android.material:material:1.0.0'
     //percent = 'com.android.support:percent:24.1.1'
 
 }
